mod: tweak dtv bots, waves & add units (#1162)

+++ src/Module.Server/Modes/Dtv/CrpgDtvData.cs
@@ -31,5 +31,5 @@ public class CrpgDtvGroup
     public string ClassDivisionId { get; set; } = string.Empty;
 
     [XmlAttribute(AttributeName = "count")]
-    public int Count { get; set; }
+    public float Count { get; set; }
 }
